% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RQUIS.R
\name{quis.getEngine}
\alias{quis.getEngine}
\title{Get an engine instance.}
\usage{
quis.getEngine()
}
\value{
An instance of the query execution engine is returned so that it is possible for the following functions to interact with
the engine. It is possible to have more than one engine instances in a single R script and use them for different processing porpuses.
}
\description{
\code{quis.getEngine} instantiates a new query execution engine, so that a user can submit queries to and ask for execution.
}
\details{
The function uses Java Development Kit (JDK) 8 or upper. Having the Java Runtime Engine (JRE) is not enough.
Also the package needs an environment variable named JAVA_HOME that points to the root folder of the JDK
For more information consult the ReadMe file in the package root folder.
}
\examples{
engine <- quis.getEngine()

\dontrun{
engine <- quis.getEngine()
}
}

