% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RQUIS.R
\name{quis.getProcess}
\alias{quis.getProcess}
\title{Returns the process.}
\usage{
quis.getProcess(engine)
}
\arguments{
\item{engine}{the engine instance created by \code{quis.getEngine}}
}
\value{
the process submitted to the engine
}
\description{
\code{quis.getProcess} Returns the process that was submitted to an engine
using \code{quis.addScript} or \code{quis.loadProcess}.
}
\examples{
engine1 <- quis.getEngine()
addedScript <- quis.loadProcess(engine1, system.file("extdata", "ex2.xqt", package="RQUIS"))
process <- quis.getProcess(engine1)

\dontrun{
engine1 <- quis.getEngine()
addedScript <- quis.loadProcess(engine1, system.file("extdata", "ex2.xqt", package="RQUIS"))
process <- quis.getProcess(engine1)
}
}

