% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RQUIS.R
\name{quis.getRunReport}
\alias{quis.getRunReport}
\title{Reports the result of the execution.}
\usage{
quis.getRunReport(engine)
}
\arguments{
\item{engine}{the engine instance created by \code{quis.getEngine}}
}
\value{
A verbose report explaining the execution of each statement including the overall execution time, result size, and errors.
}
\description{
\code{quis.getRunReport} Reports whether execution of the proceess has encountered any error.
}
\examples{
engine1 <- quis.getEngine()
addedScript <- quis.loadProcess(engine1, system.file("extdata", "ex2.xqt", package="RQUIS"))
quis.runProcess(engine1)
err <- quis.getRunReport(engine1)
\dontrun{
engine1 <- quis.getEngine()
addedScript <- quis.loadProcess(engine1, system.file("extdata", "ex2.xqt", package="RQUIS"))
quis.runProcess(engine1)
err <- quis.getRunReport(engine1)
}
}

