% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RQUIS.R
\name{quis.getVariable}
\alias{quis.getVariable}
\title{Gets the resultset by variable name.}
\usage{
quis.getVariable(engine, variableName)
}
\arguments{
\item{engine}{the engine instance created by \code{quis.getEngine}}

\item{variableName}{the variable name that is used as the target of a query in the process.}
}
\value{
A data frame equivalent to the result set of the corrsponding query as in the process.
}
\description{
\code{quis.getVariable} Gets the result set, pointed to by a variable, as a data frame.
}
\examples{
engine1 <- quis.getEngine()
quis.loadProcess(engine1, system.file("extdata", "ex2.xqt", package="RQUIS"))
quis.runProcess(engine1)
dlm <- quis.getVariable(engine1, "meanDailyTemp")

\dontrun{
engine1 <- quis.getEngine()
quis.loadProcess(engine1, system.file("extdata", "ex2.xqt", package="RQUIS"))
quis.runProcess(engine1)
dlm <- quis.getVariable(engine1, "meanDailyTemp")
}
}

