% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RQUIS.R
\name{quis.listVariables}
\alias{quis.listVariables}
\title{Lists the name of the variables that contain tabular data.}
\usage{
quis.listVariables(engine)
}
\arguments{
\item{engine}{the engine instance created by \code{quis.getEngine}}
}
\description{
\code{quis.listVariables} retrieves the name of the variables that contain tabular data.
}
\examples{
engine1 <- quis.getEngine()
quis.loadProcess(engine1, system.file("extdata", "ex2.xqt", package="RQUIS"))
quis.runProcess(engine1)
varNames <- quis.listVariables(engine1)

\dontrun{
engine1 <- quis.getEngine()
quis.loadProcess(engine1, system.file("extdata", "ex2.xqt", package="RQUIS"))
quis.runProcess(engine1)
varNames <- quis.listVariables(engine1)
}
}

