% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RQUIS.R
\name{quis.loadProcess}
\alias{quis.loadProcess}
\title{Loads a process file into the engine.}
\usage{
quis.loadProcess(engine, fileName)
}
\arguments{
\item{engine}{the engine instance created by \code{quis.getEngine}}

\item{fileName}{the full path to a file containing the process}
}
\value{
the content of the file loaded
}
\description{
\code{quis.loadProcess} Loads a process stored in a file into the engine.
}
\examples{
engine1 <- quis.getEngine()
addedScript <- quis.loadProcess(engine1, system.file("extdata", "ex2.xqt", package="RQUIS"))

\dontrun{
engine1 <- quis.getEngine()
addedScript <- quis.loadProcess(engine1, system.file("extdata", "ex2.xqt", package="RQUIS"))
}
}

