% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RQUIS.R
\name{quis.runProcess}
\alias{quis.runProcess}
\title{Runs the process.}
\usage{
quis.runProcess(engine)
}
\arguments{
\item{engine}{the engine instance created by \code{quis.getEngine}}
}
\value{
OK if everything went ok! To be sure that no error has happend use \code{quis.getRunReport}.
The result set of submitted queries are not returned automatically. They should be requested for using \code{quis.getVariabe}
}
\description{
\code{quis.runProcess} Runs the submitted process and keeps the result sets for later use.
}
\examples{
engine1 <- quis.getEngine()
addedScript <- quis.loadProcess(engine1, system.file("extdata", "ex2.xqt", package="RQUIS"))
quis.runProcess(engine1)

\dontrun{
engine1 <- quis.getEngine()
addedScript <- quis.loadProcess(engine1, system.file("extdata", "ex2.xqt", package="RQUIS"))
quis.runProcess(engine1)
}
}

