% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RQUIS.R
\name{quis.versionInfo}
\alias{quis.versionInfo}
\title{Query engine information.}
\usage{
quis.versionInfo()
}
\description{
\code{quis.versionInfo} returns the name of the query engine, its version and the maturity level.
}

