% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RQt.R
\name{rqt.addScript}
\alias{rqt.addScript}
\title{Adds one or more statements to the current process.}
\usage{
rqt.addScript(engine, statement)
}
\arguments{
\item{engine}{the engine instance created by \code{rqt.getEngine}}

\item{statement}{the statement to be submitted to the engine for later execution. Statements are of one of these types:
PERSPECTIVE, CONNECTION, BINDING, SELECT. A minimum process needs at least one connection. one binding, and one select.
It is possible to have any number of these statements, but they have to follow the abovementioned order,
which means e.g., all perspectives go before all connections and so on.}
}
\value{
the added script is returned. Usually the return value is not needed, but in case of debugging it may be useful.
}
\description{
\code{rqt.addScript} Adds one or more statements to the current process. Each engine instance works on one process at a time.
}
\examples{
engine1 <- rqt.getEngine()
addedScript <- rqt.addScript(engine1, "CONNECTION cnn1 ADAPTER=CSV SOURCE_URI='data\\\\\\\\' PARAMETERS=delimiter:comma, fileExtension:csv, firstRowIsHeader:true")

\dontrun{
engine1 <- rqt.getEngine()
addedScript <- rqt.addScript(engine1, "CONNECTION cnn1 ADAPTER=CSV SOURCE_URI='data\\\\\\\\' PARAMETERS=delimiter:comma, fileExtension:csv, firstRowIsHeader:true")
}
}

