% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RQt.R
\name{rqt.getPlot}
\alias{rqt.getPlot}
\title{Gets the visual resultset by variable name.}
\usage{
rqt.getPlot(engine, plotName)
}
\arguments{
\item{engine}{the engine instance created by \code{rqt.getEngine}}

\item{plotName}{the plot name as used in the process as the target of a query.}
}
\value{
An image of type JPeg as rendered by the corrsponding query in the submitted process.
}
\description{
\code{rqt.getVariable} Gets the visual result set, pointed to by a variable, as an image.
}
\examples{
engine1 <- rqt.getEngine()
addedScript <- rqt.loadProcess(engine1, "Examples\\\\processes\\\\ex2.xqt")
rqt.runProcess(engine1)
dlm <- rqt.getPlot(engine1, "p2")
\dontrun{
engine1 <- rqt.getEngine()
addedScript <- rqt.loadProcess(engine1, "Examples\\\\processes\\\\ex2.xqt")
rqt.runProcess(engine1)
dlm <- rqt.getPlot(engine1, "p2")
}
}

