% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RQt.R
\name{rqt.getProcess}
\alias{rqt.getProcess}
\title{Returns the process.}
\usage{
rqt.getProcess(engine)
}
\arguments{
\item{engine}{the engine instance created by \code{rqt.getEngine}}
}
\value{
the process submitted to the engine
}
\description{
\code{rqt.getProcess} Returns the process that is submitted to the engine either using \code{rqt.addScript} or \code{rqt.loadProcess}.
}
\examples{
engine1 <- rqt.getEngine()
addedScript <- rqt.loadProcess(engine1, "Examples\\\\processes\\\\ex2.xqt")
process <- rqt.getProcess(engine1)

\dontrun{
engine1 <- rqt.getEngine()
addedScript <- rqt.loadProcess(engine1, "Examples\\\\processes\\\\ex2.xqt")
process <- rqt.getProcess(engine1)
}
}

