% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RQt.R
\name{rqt.getRunReport}
\alias{rqt.getRunReport}
\title{Reports the execution.}
\usage{
rqt.getRunReport(engine)
}
\arguments{
\item{engine}{the engine instance created by \code{rqt.getEngine}}
}
\value{
A verbose report explaining the execution of each statement including the execution time, result size, and errors.
}
\description{
\code{rqt.getRunReport} Reports whether execution of the proceess has encountered any error.
}
\examples{
engine1 <- rqt.getEngine()
addedScript <- rqt.loadProcess(engine1, "Examples\\\\processes\\\\ex2.xqt")
rqt.runProcess(engine1)
err <- rqt.getRunReport(engine1)
\dontrun{
engine1 <- rqt.getEngine()
addedScript <- rqt.loadProcess(engine1, "Examples\\\\processes\\\\ex2.xqt")
rqt.runProcess(engine1)
err <- rqt.getRunReport(engine1)
}
}

