% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RQt.R
\name{rqt.getVariableSchema}
\alias{rqt.getVariableSchema}
\title{Gets the schema of a variable.}
\usage{
rqt.getVariableSchema(engine, variableName, extended = FALSE)
}
\arguments{
\item{engine}{the engine instance created by \code{rqt.getEngine}}

\item{variableName}{the variable name.}

\item{extended}{determine whether simple or semantically enhanced schema should be returned.}
}
\value{
A data frame that contains name, type, constaints, and semantic annotation of the attributes of the target variable's schema.
}
\description{
\code{rqt.getVariableSchema} Gets the schema of a variable, as a data frame.
}
\examples{
engine1 <- rqt.getEngine()
rqt.loadProcess(engine1, system.file("extdata", "ex2.xqt", package="RQt"))
rqt.runProcess(engine1)
scm <- rqt.getVariableSchema(engine1, "meanDailyTemp")

\dontrun{
engine1 <- rqt.getEngine()
rqt.loadProcess(engine1, system.file("extdata", "ex2.xqt", package="RQt"))
rqt.runProcess(engine1)
scm <- rqt.getVariableSchema(engine1, "meanDailyTemp")
}
}

