% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RQt.R
\name{rqt.listVariables}
\alias{rqt.listVariables}
\title{Lists the name of the variables that contain tabular data.}
\usage{
rqt.listVariables(engine)
}
\arguments{
\item{engine}{the engine instance created by \code{rqt.getEngine}}
}
\description{
\code{rqt.listVariables} retrieves the name of the variables that contain tabular data.
}
\examples{
engine1 <- rqt.getEngine()
rqt.loadProcess(engine1, system.file("extdata", "ex2.xqt", package="RQt"))
rqt.runProcess(engine1)
varNames <- rqt.listVariables(engine1)

\dontrun{
engine1 <- rqt.getEngine()
rqt.loadProcess(engine1, system.file("extdata", "ex2.xqt", package="RQt"))
rqt.runProcess(engine1)
varNames <- rqt.listVariables(engine1)
}
}

