% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RQt.R
\name{rqt.loadProcess}
\alias{rqt.loadProcess}
\title{Loads a process file into the engine.}
\usage{
rqt.loadProcess(engine, fileName)
}
\arguments{
\item{engine}{the engine instance created by \code{rqt.getEngine}}

\item{fileName}{the full path to a file containing the process}
}
\value{
the content of the file loaded
}
\description{
\code{rqt.loadProcess} Loads a process stored in a file into the engine.
}
\examples{
engine1 <- rqt.getEngine()
addedScript <- rqt.loadProcess(engine1, system.file("Examples/processes", "ex2.xqt", package="RQt"))

\dontrun{
engine1 <- rqt.getEngine()
addedScript <- rqt.loadProcess(engine1, system.file("Examples/processes", "ex2.xqt", package="RQt"))
}
}

