% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RQt.R
\name{rqt.runProcess}
\alias{rqt.runProcess}
\title{Runs the process.}
\usage{
rqt.runProcess(engine)
}
\arguments{
\item{engine}{the engine instance created by \code{rqt.getEngine}}
}
\value{
OK if everything went ok! To be sure that no error has happend use \code{rqt.getRunReport}.
The result set of submitted queries are not returned automatically. They should be requested for using \code{rqt.getVariabe}
}
\description{
\code{rqt.runProcess} Runs the submitted process and keeps the result sets for later use.
}
\examples{
engine1 <- rqt.getEngine()
addedScript <- rqt.loadProcess(engine1, system.file("extdata", "ex2.xqt", package="RQt"))
rqt.runProcess(engine1)

\dontrun{
engine1 <- rqt.getEngine()
addedScript <- rqt.loadProcess(engine1, system.file("extdata", "ex2.xqt", package="RQt"))
rqt.runProcess(engine1)
}
}

