% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RQt.R
\name{rqt.versionInfo}
\alias{rqt.versionInfo}
\title{Query engine information.}
\usage{
rqt.versionInfo()
}
\description{
\code{rqt.versionInfo} returns the name of the query engine, its version and the maturity level.
}

