% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RQt.R
\name{rqt.getVariable}
\alias{rqt.getVariable}
\title{Gets the resultset by variable name.}
\usage{
rqt.getVariable(engine, variableName)
}
\arguments{
\item{engine}{the engine instance created by \code{rqt.getEngine}}

\item{variableName}{the variable name tht is used as the target of a query in the process.}
}
\value{
A data frame equivalent to the result set of the corrsponding query as in the process.
}
\description{
\code{rqt.getVariable} Gets the result set, pointed to by a variable, as a data frame.
}
\examples{
engine1 <- rqt.getEngine()
rqt.loadProcess(engine1, system.file("extdata", "ex2.xqt", package="RQt"))
rqt.runProcess(engine1)
dlm <- rqt.getVariable(engine1, "meanDailyTemp")

\dontrun{
engine1 <- rqt.getEngine()
rqt.loadProcess(engine1, system.file("extdata", "ex2.xqt", package="RQt"))
rqt.runProcess(engine1)
dlm <- rqt.getVariable(engine1, "meanDailyTemp")
}
}

